requirements.txt
Django==5.1.3
psycopg2-binary==2.9.10 
pillow==11.0.0 